# docs/openapi.yaml
openapi: 3.0.3
info:
  title: NIAC-Go API
  description: |
    Network In A Can - Go Edition REST API

    This API provides programmatic access to network simulation management,
    device monitoring, configuration updates, and PCAP replay functionality.

    ## Authentication
    All endpoints require Bearer token authentication. Set the token via
    the `NIAC_API_TOKEN` environment variable when starting the server.

    ## CSRF Protection
    State-changing operations (POST, PUT, PATCH, DELETE) require a CSRF token.
    1. Retrieve token from `/api/v1/csrf-token`
    2. Include token in `X-CSRF-Token` header

    ## Rate Limiting
    - Default: 100 requests/second per IP
    - Burst: 200 requests
    - Response: HTTP 429 when exceeded
  version: 2.8.1
  contact:
    name: Kris Armstrong
    url: https://github.com/MustardSeedNetworks/niac-go
  license:
    name: MIT
    url: https://github.com/MustardSeedNetworks/niac-go/blob/main/LICENSE
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://niac.example.com/api/v1
    description: Production server (example)
security:
  - BearerAuth: []
...
paths:
...
  /devices:
    get:
      tags:
        - Devices
      summary: List all devices
      description: Returns list of all configured simulated devices
      operationId: listDevices
      responses:
        '200':
          description: Device list
          content:
            application/json:
              schema:
                type: object
                properties:
                  devices:
                    type: array
                    items:
                      $ref: '#/components/schemas/Device'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /devices/{name}:
    get:
      tags:
        - Devices
      summary: Get device details
      description: Returns detailed information about a specific device
      operationId: getDevice
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
          description: Device name
          example: "router-1"
      responses:
        '200':
          description: Device details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Token
      description: API token set via NIAC_API_TOKEN environment variable

    CsrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: CSRF token obtained from /api/v1/csrf-token
  schemas:
...
    Device:
      type: object
      properties:
        name:
          type: string
          example: "router-1"
        type:
          type: string
          enum: [router, switch, server, firewall, loadbalancer]
          example: "router"
        mac:
          type: string
          example: "00:0c:29:ab:cd:ef"
        ip:
          type: string
          example: "192.168.1.1"
        ip_addresses:
          type: array
          items:
            type: string
          example: ["192.168.1.1", "10.0.0.1"]
        gateway:
          type: string
          example: "192.168.1.254"
        protocols:
          type: array
          items:
            type: string
          example: ["ARP", "ICMP", "DNS", "SNMP", "LLDP"]
        state:
          type: string
          enum: [up, down]
          example: "up"
        stats:
          type: object
          properties:
            packets_sent:
              type: integer
              format: int64
            packets_received:
              type: integer
              format: int64
...
    Error:
      type: object
      properties:
        error_code:
          type: string
          example: "invalid_request"
        message:
          type: string
          example: "Invalid request parameters"
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: "speed"
              issue:
                type: string
                example: "must be between 0.1 and 10.0"
        request_id:
          type: string
          description: Unique request ID for tracing
          example: "abc123def456789"
  responses:
...
    Unauthorized:
      description: Unauthorized - missing or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error_code: "unauthorized"
            message: "Invalid or missing authentication token"
            request_id: "abc123def456789"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error_code: "not_found"
            message: "Device not found"
            request_id: "abc123def456789"
...